// apps/cms/src/shared/api/generated/members/schemas/paymentStatementBasicInfo.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * JA Korea CMS Backend API — Members subset
 * Filtered for CMS member management Orval codegen.
 * OpenAPI spec version: v9
 */

/**
 * 지급조서 사전 동의서의 작성 폼 외부 기본정보 sidecar
 */
export interface PaymentStatementBasicInfo {
  /**
     * @minLength 0
     * @maxLength 100
     */
  nameKo?: string;
  /**
     * @minLength 0
     * @maxLength 100
     */
  nameEn?: string;
  /**
     * @minLength 0
     * @maxLength 6
     */
  residentFront?: string;
  /**
     * @minLength 0
     * @maxLength 7
     */
  residentBack?: string;
  /**
     * @minLength 0
     * @maxLength 200
     */
  affiliation?: string;
  noAffiliation?: boolean;
  /**
     * @minLength 0
     * @maxLength 500
     */
  addressRoad?: string;
  /**
     * @minLength 0
     * @maxLength 500
     */
  addressDetail?: string;
  /**
     * @minLength 0
     * @maxLength 100
     */
  bankName?: string;
  /**
     * @minLength 0
     * @maxLength 100
     */
  accountNumber?: string;
  /**
     * @minLength 0
     * @maxLength 100
     */
  accountHolder?: string;
  /**
     * @minLength 0
     * @maxLength 200
     */
  paymentPurpose?: string;
}